requirements.txt
langchain
langchain-community
langchain_google_genai
python-dotenv
faiss-cpu
networkx
PyPDF2
python-pptx
pypdf
langchain-text-splitters